Replaced wget with custom serverspec resource
